+++ R/fitted.R
@@ -16,7 +16,7 @@ fitted.mdl_df <- function(object, ...){
   kv <- key_vars(object)
   object <- mutate(as_tibble(object), 
               dplyr::across(all_of(mbl_vars), function(x) lapply(x, fitted, ...)))
-  object <- pivot_longer(object, mbl_vars, names_to = ".model", values_to = ".fitted")
+  object <- pivot_longer(object, all_of(mbl_vars), names_to = ".model", values_to = ".fitted")
   unnest_tsbl(object, ".fitted", parent_key = c(kv, ".model"))
 }