+++ R/residuals.R
@@ -13,7 +13,7 @@ residuals.mdl_df <- function(object, ...){
   kv <- key_vars(object)
   object <- mutate(as_tibble(object), 
               dplyr::across(all_of(mbl_vars), function(x) lapply(x, residuals, ...)))
-  object <- pivot_longer(object, mbl_vars, names_to = ".model", values_to = ".resid")
+  object <- pivot_longer(object, all_of(mbl_vars), names_to = ".model", values_to = ".resid")
   unnest_tsbl(object, ".resid", parent_key = c(kv, ".model"))
 }