+++ NEWS.md
@@ -1,4 +1,21 @@
-# fabletools (development version)
+# fabletools 0.4.1
+
+Minor patch for upcoming release of ggdist v3.3.1
+
+## Improvements
+
+* Added (scaled) pinball loss metrics to `interval_accuracy_measures` (#379).
+* Improved use of random seed in parallel modelling and forecasting (#384).
+* Documentation improvements
+
+# fabletools 0.4.0
+
+## Improvements
+
+* Improved handling of `combination_model()` when used with transformed
+  component models.
+* `autoplot(<fbl_ts>)`, `autolayer(<fbl_ts>)` and `autoplot(<dcmp_ts>)` now use
+  the ggdist package visualising uncertainty with distributional vectors.
 
 # fabletools 0.3.4